requirements.txt
Flask>=2.3,<4
Flask-SQLAlchemy>=3.1,<4
Flask-Login>=0.6,<1
Flask-Migrate>=4.0,<5
python-dotenv>=1.0,<2
pytest>=7,<9
